requirements.txt
a2wsgi==1.10.10
attr==0.3.2
beautifulsoup4==4.14.3
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==46.0.5
docutils==0.22.4
fastapi==0.135.1
filelock==3.25.2
firecrawl==4.19.0
groq==1.1.1
gunicorn==25.1.0
HTMLParser==0.0.2
httptools==0.7.1
ipython==8.12.3
ipywidgets==8.1.8
jnius==1.1.0
joblib==1.5.3
keyring==25.7.0
langdetect==1.0.9
mock==5.2.0
numpy==2.4.3
peft==0.18.1
Pillow==12.1.1
protobuf==7.34.1
pydantic==2.12.5
pyOpenSSL==26.0.0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
scikit_learn==1.8.0
Sphinx==9.1.0
thread==2.0.6
torch==2.10.0
transformers==5.3.0
twilio==9.10.3
urllib3_secure_extra==0.1.0
watchfiles==1.1.1
wsproto==1.3.2
xmlrpclib==1.0.1
